Update project metafiles for recent changes to the GDK build system:
Use the new build target for SDK level 19. Previously Glass was built on SDK
level 14 (v4.0.3).

Remove the declared triggers to turn on Wi-Fi and Bluetooth, and for turning off
airplane mode. These can be done from the settings bundle with the touchpad, and
aren't really the kind of things I need on a quick command. I've left the string
resources for the triggers, so you can add them back if you want them.

Add the development permission to let this package add the triggers to the OK
Glass menu. The triggers will not be registered unless this is declared.